requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Mail==0.9.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
typing_extensions==4.11.0
Werkzeug==3.0.2
